requirements.txt
openai_whisper==20240930
protobuf==5.29.3
streamlit==1.41.1
